servidor iniciado, se realizan cambios que no lo dejaban iniciar
