+++ backend/src/inspections/enums/form-type.enum.ts
@@ -1,0 +1,6 @@
+export enum FormType {
+  WORK_AREAS = 'areas y puestos de trabajo',
+  MACHINERY = 'maquinaria y equipos',
+  PROTECTIVE_EQUIPMENT = 'elementos de protección personal',
+  MEDICAL = 'botiquines y camillas',
+}
